Refactor: models, users service, users controller 분리
