+++ hyunjin/middlewares/index.js
@@ -1,0 +1,15 @@
+exports.throwError = (code, message) => {
+  if (!code) return;
+  const error = new Error();
+  let errorMessage = new Map([
+    [400, 'bad request'],
+    [401, 'unAuthorized'],
+    [500, 'internal server error'],
+  ]);
+  if (!errorMessage.get(code) || message) {
+    errorMessage.set(code, message);
+  }
+  error.message = errorMessage.get(code);
+  error.status = code;
+  throw error;
+};
